// src/types/contracts/protocol/interfaces/IDolomiteMargin.ts
export declare namespace IDolomiteStructs {
  export type AccountInfoStruct = { owner: string; number: BigNumberish };
// ...
}